:recycle: Split up projects schema
